refactor: modify float to double for product entity

+++ product-service/src/main/java/com/nadiannis/product_service/entity/Product.java
@@ -25,7 +25,7 @@ public class Product {
 
     private String name;
 
-    private Float price;
+    private Double price;
 
     private String category;
